requirements.txt
asgiref==3.5.2
Django==3.2.10
django-environ==0.9.0
djangorestframework==3.11.2
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2

